requirements.txt
hvac
boto3
python-gnupg
pyyaml
